ci: optimize build lib
