--- docs/vite.config.ts
@@ -1,13 +1,0 @@
-import path from 'path'
-import { defineConfig } from 'vite'
-import plugins from './src/plugin/common-plugins'
-
-export default defineConfig({
-  base: '/element-pro-components/',
-  resolve: {
-    alias: {
-      '/@src': path.resolve(__dirname, '../src'),
-    },
-  },
-  plugins,
-})
