+++ package.json
@@ -18,11 +18,12 @@
   },
   "scripts": {
     "dev": "vite",
-    "build": "vite build --emptyOutDir && npm run build:helper && npm run build:styles && npm run build:types",
+    "build": "npm run build:lib && npm run build:helper && npm run build:styles && npm run build:types",
+    "build:lib": "vite build -c vite.build.config.ts --emptyOutDir",
     "build:types": "rimraf types && tsc -d",
     "build:helper": "node .helperrc.js",
     "build:styles": "postcss src/styles/*.css -c .postcssrc.json -d lib/styles",
-    "build:docs": "vite build docs --ssrManifest --outDir dist/static && vite build docs --ssr src/entry-server.ts --outDir dist/server && node docs/prerender.js",
+    "build:docs": "vite-ssg build",
     "test": "jest",
     "lint": "eslint . --fix --ext .ts,.vue,.js --ignore-pattern \"/lib/\" --ignore-pattern \"/types/\"",
     "prettier": "prettier --check --write --ignore-unknown \"{example,src,docs,test}/**\""
@@ -69,6 +70,7 @@
     "@vue/compiler-sfc": "^3.1.4",
     "@vue/server-renderer": "^3.1.4",
     "@vue/test-utils": "^2.0.0-rc.10",
+    "@vueuse/head": "^0.6.0",
     "babel-jest": "^26.6.3",
     "components-helper": "^1.0.3",
     "escape-html": "^1.0.3",
@@ -92,7 +94,7 @@
     "ts-jest": "^26.5.5",
     "typescript": "^4.2.4",
     "vite": "^2.4.1",
-    "vite-plugin-md": "^0.6.7",
+    "vite-plugin-md": "^0.10.0",
     "vite-plugin-pwa": "^0.8.2",
     "vue-jest": "^5.0.0-alpha.10"
   },
